refacto : move AlreadyAssociatedException to Exception folder root
