+++ src/main/java/com/jeyofdev/yellow_berry/domain/wishlist/WishlistService.java
@@ -7,9 +7,8 @@ import com.jeyofdev.yellow_berry.domain.product.Product;
 import com.jeyofdev.yellow_berry.domain.product.ProductRepository;
 import com.jeyofdev.yellow_berry.domain.profile.Profile;
 import com.jeyofdev.yellow_berry.domain.profile.ProfileService;
-import com.jeyofdev.yellow_berry.domain.tag.Tag;
 import com.jeyofdev.yellow_berry.exception.AlreadyTakenException;
-import com.jeyofdev.yellow_berry.exception.model.AlreadyAssociatedException;
+import com.jeyofdev.yellow_berry.exception.AlreadyAssociatedException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
